requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1

# Financial data
yfinance==0.2.28

# Machine learning
scikit-learn==1.3.0

# Utilities
pyyaml==6.0.1
tqdm==4.65.0

# Jupyter support (optional)
jupyter==1.0.0
notebook==7.0.2
ipywidgets==8.1.0

# Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0

# Code quality (optional)
black==23.7.0
flake8==6.1.0
mypy==1.4.1